requirements.txt
flask
nltk
spacy
google-cloud-translate
